ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
groq
pdfplumber
pandas
sentence-transformers
faiss-cpu
scikit-learn
numpy
